refactor(GameServer): pattern match the game_won function
Previously a cond was used with a series of checks on the grid position
for the same piece. This has been changed to pattern match using a
representation of the game grid.

The draw condition is inside of a function as it is not possible to use
`in` with a dynamic list inside a pattern match.